requirements.txt
pydicom==2.4.3
numpy==1.26.0
pyinstaller==6.0.0
pythonocc-core==7.7.2
PySide6==6.5.2